requirements.txt
git+https://github.com/Pycord-Development/pycord
mysql.connector